Split lower-third settings into its own module
Extracts Lower Third settings UI and view model from :composeApp into a new `:lower-third-settings-tab` module, including its tests, screenshots, detekt baseline, and module docs (`AGENT.md`/`CLAUDE.md`).

Wires the new module into Gradle, settings, app dependencies, `test-changed.sh`, and CI/workflows for detekt, tests, coverage, screenshot recording, and report publishing. Also introduces shared lower-third test fixtures in `:lower-third-tab` testFixtures so folder-scan/Lottie fixtures are reused across modules, updates server-tab tests to consume them, and separates song settings view-model coverage into a dedicated `SongSettingsViewModelTest` in `:composeApp`.
