+++ .github/workflows/test.yml
@@ -216,6 +216,15 @@ jobs:
               - 'diagnostics/**'          # implementation(projects.diagnostics)
               - 'song-chords/**'          # implementation(projects.songChords)
               - 'presentation-engine/**'  # implementation(projects.presentationEngine)
+            lowerthirdsettingstab:
+              - *shared
+              - 'lower-third-settings-tab/**'
+              - 'lower-third-tab/**'      # implementation — LottieFonts, and the shared Lottie fixtures
+              - 'companion-server/**'     # via :lower-third-tab's api — isLottieFile
+              - 'settings/**'             # api(projects.settings)
+              - 'ui-components/**'        # implementation(projects.uiComponents)
+              - 'theme/**'                # testImplementation(projects.theme)
+              - 'resources/**'            # api(projects.resources)
 
       - name: Set up JDK 21
         uses: actions/setup-java@v4
@@ -253,6 +262,7 @@ jobs:
           :statistics:detekt
           :ui-components:detekt :dictionary-tab:detekt :dictionary-settings-tab:detekt
           :announcements-tab:detekt :lower-third-tab:detekt
+          :lower-third-settings-tab:detekt
 
       - name: App unit tests
         id: unit_tests
@@ -454,6 +464,14 @@ jobs:
         if: always() && steps.modules.outputs.lowerthirdtab == 'true'
         run: ./gradlew :lower-third-tab:jacocoTestCoverageVerification -x test
 
+      - name: Lower Third settings tab tests
+        if: always() && steps.modules.outputs.lowerthirdsettingstab == 'true'
+        run: ./gradlew :lower-third-settings-tab:test
+
+      - name: Lower Third settings tab coverage floor
+        if: always() && steps.modules.outputs.lowerthirdsettingstab == 'true'
+        run: ./gradlew :lower-third-settings-tab:jacocoTestCoverageVerification -x test
+
       - name: Theme tests
         if: always() && steps.modules.outputs.theme == 'true'
         run: ./gradlew :theme:test
@@ -609,6 +627,7 @@ jobs:
             dictionary-settings-tab/build/test-results/**/TEST-*.xml
             announcements-tab/build/test-results/**/TEST-*.xml
             lower-third-tab/build/test-results/**/TEST-*.xml
+            lower-third-settings-tab/build/test-results/**/TEST-*.xml
             companion-server/build/test-results/**/TEST-*.xml
           # Paired with report_paths BY INDEX -- keep the two lists in the same order, and the same
           # LENGTH. A merge once left eight paths against six names, which does not fail: it shifts
@@ -639,6 +658,7 @@ jobs:
             Dictionary settings tab
             Announcements tab
             Lower Third tab
+            Lower Third settings tab
             Companion Server
           detailed_summary: true
           include_passed: false
@@ -802,5 +822,7 @@ jobs:
             announcements-tab/build/test-results/
             lower-third-tab/build/reports/tests/
             lower-third-tab/build/test-results/
+            lower-third-settings-tab/build/reports/tests/
+            lower-third-settings-tab/build/test-results/
             companion-server/build/reports/tests/
             companion-server/build/test-results/
