+++ composeApp/build.gradle.kts
@@ -333,6 +333,7 @@ kotlin {
             implementation(projects.dictionaryTab)
             implementation(projects.dictionarySettingsTab)
             implementation(projects.lowerThirdTab)
+            implementation(projects.lowerThirdSettingsTab)
             // The HTTP/WebSocket surface the app exposes: the wire format, the routes, TLS, the
             // tunnel and the instance-link client. What a remote request then does to the app is
             // `remote/` here.
@@ -492,6 +493,11 @@ dependencies {
     // app's remote-command and instance-link suites drive a real server, so they borrow the fixture
     // from the module that owns the wire format rather than standing one up twice.
     add("jvmTestImplementation", testFixtures(projects.companionServer))
+    // The Lottie-folder fixtures -- what counts as an animation on disk, and the wait for the
+    // scan that decides. The server tab's lower-third triggers are driven from a real folder
+    // here, and the options page that lists the same folder now lives in
+    // :lower-third-settings-tab, so both read one definition rather than a copy each.
+    add("jvmTestImplementation", testFixtures(projects.lowerThirdTab))
 }
 
 compose.desktop {
