+++ composeApp/src/jvmTest/kotlin/org/churchpresenter/app/churchpresenter/dialogs/tabs/ServerSettingsTabTestSupport.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.performTextReplacement
 import androidx.compose.ui.test.runComposeUiTest
 import org.churchpresenter.app.churchpresenter.data.RemoteClientManager
+import org.churchpresenter.lowerthird.awaitFolderScan
 import org.churchpresenter.settings.AppSettings
 import org.churchpresenter.settings.ServerSettings
 import org.churchpresenter.companionserver.CompanionServer
@@ -237,3 +238,24 @@ internal fun ComposeUiTest.boxShowing(value: String): SemanticsNodeInteraction =
 /** How many times [text] is rendered. */
 internal fun ComposeUiTest.countOf(text: String): Int =
     onAllNodesWithText(text).fetchSemanticsNodes(atLeastOneRootRequired = false).size
+
+/**
+ * Waits for the lower-third rows that [settings]' folder should produce.
+ *
+ * The server tab loads them off the UI thread too, but unlike the lower-third tab it starts from an
+ * empty list with no "Scanning folder" marker — so `awaitFolderScan` passes vacuously here and an
+ * assertion would run against an empty list. Waiting for the names the folder actually holds is a
+ * positive signal that the read landed, and it stays correct for an empty folder (nothing to wait
+ * for).
+ */
+internal fun ComposeUiTest.awaitLowerThirdRows(settings: AppSettings) {
+    val folder = File(settings.streamingSettings.lowerThirdFolder)
+    val names = folder.listFiles()
+        ?.filter { it.extension.equals("json", ignoreCase = true) && it.readText().contains("\"layers\"") }
+        ?.map { it.nameWithoutExtension }
+        ?: return
+    if (names.isEmpty()) return
+    waitUntil("the lower-third rows loaded") {
+        names.all { onAllNodesWithText(it, substring = true).fetchSemanticsNodes(false).isNotEmpty() }
+    }
+}
