+++ lower-third-settings-tab/CLAUDE.md
@@ -1,0 +1,1 @@
+@AGENT.md
